Changes to model

+++ src/main/java/edu/csula/vkc/models/Trim.java
@@ -2,28 +2,24 @@ package edu.csula.vkc.models;
 
 public class Trim {
 
-	int trim_id;
+	long trim_id;
 	String trim;
 	double price;
 	String milage;
 	public Trim() {
 		super();
 	}
-	
-	
-	public Trim(int trim_id, String trim, double price, String milage) {
+	public Trim(long trim_id, String trim, double price, String milage) {
 		super();
 		this.trim_id = trim_id;
 		this.trim = trim;
 		this.price = price;
 		this.milage = milage;
 	}
-
-
-	public int getTrim_id() {
+	public long getTrim_id() {
 		return trim_id;
 	}
-	public void setTrim_id(int trim_id) {
+	public void setTrim_id(long trim_id) {
 		this.trim_id = trim_id;
 	}
 	public String getTrim() {
@@ -44,7 +40,6 @@ public class Trim {
 	public void setMilage(String milage) {
 		this.milage = milage;
 	}
-	
 	
 	
 }
